refactor: extract charts and tour routers (PR-11, Phase 1)
